Module contains function to generate job counts for each skill
This module reads in the clean csv of skills from payscale. 
It defines a function called "return_job_counts" that takes three arguments for "job position", "city", and "state abbreviation", and returns a sorted pandas data frame of each skill that matched at least one job, and the number of jobs that skilled matched.